+++ app/pages/clickerList/clickerList.e2e.ts
@@ -1,7 +1,7 @@
 import { ElementFinder } from 'protractor';
 
 let clickerField: ElementFinder = element(by.css('.text-input'));
-let clickerButton: ElementFinder = element.all(by.className('button')).first();
+let clickerButton: ElementFinder = element.all(by.className('button-outline')).first();
 let removeButton: ElementFinder = element.all(by.css('.button-outline-danger')).first();
 let firstClicker: ElementFinder = element.all(by.className('clickerList')).first();