+++ app/pages/clickerList/clickerList.spec.ts
@@ -1,14 +1,16 @@
 import {
   beforeEach,
   beforeEachProviders,
-  ComponentFixture,
   describe,
   expect,
   injectAsync,
   it,
+}                               from '@angular/core/testing';
+import {
+  ComponentFixture,
   TestComponentBuilder,
-}                               from 'angular2/testing';
-import { provide }              from 'angular2/core';
+}                               from '@angular/compiler/testing';
+import { provide }              from '@angular/core';
 import { ClickerList }          from './clickerList';
 import { Utils }                from '../../services/utils';
 import {
@@ -35,7 +37,7 @@ class MockClass {
 }
 
 let clickerList: ClickerList = null;
-let clickerListFixture: ComponentFixture = null;
+let clickerListFixture: ComponentFixture<ClickerList> = null;
 
 describe('ClickerList', () => {
 
@@ -51,7 +53,7 @@ describe('ClickerList', () => {
   beforeEach(injectAsync([TestComponentBuilder], (tcb: TestComponentBuilder) => {
     return tcb
       .createAsync(ClickerList)
-      .then((componentFixture: ComponentFixture) => {
+      .then((componentFixture: ComponentFixture<ClickerList>) => {
         clickerListFixture = componentFixture;
         clickerList = componentFixture.componentInstance;
         clickerListFixture.detectChanges();