+++ app/services/clickers.ts
@@ -1,6 +1,6 @@
 'use strict';
 
-import { Injectable } from 'angular2/core';
+import { Injectable } from '@angular/core';
 import { SqlStorage } from 'ionic-angular';
 import { Clicker }    from '../models/clicker';
 import { Click }      from '../models/click';