+++ app/services/utils.spec.ts
@@ -1,6 +1,10 @@
 import { Utils } from './utils';
-import { AbstractControl, Control } from 'angular2/common';
-
+import { AbstractControl, Control } from '@angular/common';
+import {
+    describe,
+    expect,
+    it,
+}                        from '@angular/core/testing';
 describe('Utils', () => {
 
   it('resets a control', () => {